+++ Code/EnginePlugins/PhysXPlugin/WorldModule/PhysXWorldModule.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <Core/Interfaces/PhysicsWorldModule.h>
+#include <Core/Messages/TriggerMessage.h>
 #include <Core/World/Declarations.h>
 #include <Core/World/WorldModule.h>
 #include <Foundation/Threading/DelegateTask.h>
@@ -77,6 +78,12 @@ private:
   void StartSimulation(const ezWorldModule::UpdateContext& context);
   void FetchResults(const ezWorldModule::UpdateContext& context);
 
+  void HandleSimulationEvents();
+
+  void UpdatePhysicsSlideAndRollReactions();
+
+  void SpawnPhysicsImpactReactions();
+
   void HandleBrokenConstraints();
   void HandleTriggerEvents();
 
@@ -86,9 +93,9 @@ private:
 
   void UpdateJoints();
 
-  physx::PxScene* m_pPxScene;
-  physx::PxControllerManager* m_pCharacterManager;
-  ezPxSimulationEventCallback* m_pSimulationEventCallback;
+  physx::PxScene* m_pPxScene = nullptr;
+  physx::PxControllerManager* m_pCharacterManager = nullptr;
+  ezPxSimulationEventCallback* m_pSimulationEventCallback = nullptr;
 
   ezUInt32 m_uiNextShapeId;
   ezDynamicArray<ezUInt32> m_FreeShapeIds;
@@ -100,7 +107,6 @@ private:
   ezDynamicArray<ezUInt8, ezAlignedAllocatorWrapper> m_ScratchMemory;
 
   ezTime m_AccumulatedTimeSinceUpdate;
-
 
   ezPxSettings m_Settings;
 
@@ -114,3 +120,58 @@ private:
 #define EZ_PX_READ_LOCK(scene) physx::PxSceneReadLock EZ_CONCAT(pxl_, EZ_SOURCE_LINE)(scene, EZ_SOURCE_FILE, EZ_SOURCE_LINE)
 
 #define EZ_PX_WRITE_LOCK(scene) physx::PxSceneWriteLock EZ_CONCAT(pxl_, EZ_SOURCE_LINE)(scene, EZ_SOURCE_FILE, EZ_SOURCE_LINE)
+
+
+class ezPxSimulationEventCallback : public PxSimulationEventCallback
+{
+public:
+  struct InteractionContact
+  {
+    ezVec3 m_vPosition;
+    ezVec3 m_vNormal;
+    ezSurfaceResource* m_pSurface;
+    ezTempHashedString m_sInteraction;
+    float m_fImpulseSqr;
+  };
+
+  struct SlidingInfo
+  {
+    bool m_bStillSliding = false;
+    bool m_bStartedSliding = false;
+
+    bool m_bStillRolling = false;
+    bool m_bStartedRolling = false;
+
+    ezVec3 m_vPosition;
+    ezVec3 m_vNormal;
+    ezGameObjectHandle m_hSlidePrefab;
+    ezGameObjectHandle m_hRollPrefab;
+  };
+
+  ezDynamicArray<InteractionContact> m_InteractionContacts;
+  ezMap<PxRigidDynamic*, SlidingInfo> m_SlidingActors;
+  ezDeque<PxConstraint*> m_BrokenConstraints;
+
+  virtual void onConstraintBreak(PxConstraintInfo* constraints, PxU32 count) override;
+
+  virtual void onContact(const PxContactPairHeader& pairHeader, const PxContactPair* pairs, PxU32 nbPairs) override;
+
+  void SendContactReport(const PxContactPairHeader& pairHeader, const PxContactPair* pairs, PxU32 nbPairs);
+
+  struct TriggerEvent
+  {
+    ezComponentHandle m_hTriggerComponent;
+    ezComponentHandle m_hOtherComponent;
+    ezTriggerState::Enum m_TriggerState;
+  };
+
+  ezDeque<TriggerEvent> m_TriggerEvents;
+
+  virtual void onTrigger(PxTriggerPair* pairs, PxU32 count) override;
+
+  virtual void onWake(PxActor** actors, PxU32 count) override {}
+
+  virtual void onSleep(PxActor** actors, PxU32 count) override {}
+
+  virtual void onAdvance(const PxRigidBody* const* bodyBuffer, const PxTransform* poseBuffer, const PxU32 count) override {}
+};
